Update main.rs
This is the most up to date code, this implements the neural network with gradient descent with momentum. All activations are sigmoid. It also does full back propagation using Stochastic Gradient Descent with momentum, and we are doing this with sum of squared loss. We will test this with a Kaggle dataset to ensure our model works properly.